src/requirements.txt
fastapi>=0.121.2
uvicorn[standard]==0.29.0
gunicorn==23.0.0
azure-identity==1.19.0
aiohttp==3.13.1
azure-ai-inference==1.0.0b9
azure-ai-projects>=2.0.0b1
azure-search-documents

azure-core==1.34.0  # other versions might not compatible
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry==1.6.9 # version such as 1.6.11 isn't compatible
azure-search-documents
opentelemetry-sdk
setuptools==80.9.0
starlette>=0.47.2 # fix GHSA-2c2j-9gv5-cj73 (CVE-2025-54121) - DoS when parsing large multipart forms
jinja2 # new dependent of fastapi